requirements.txt
openai>=1.0.0
huggingface_hub>=0.20.0
duckduckgo-search>=6.0.0
pdfplumber>=0.10.0
python-docx>=1.0.0
click>=8.0.0
rich>=13.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0
lxml>=5.0.0
gradio>=5.0.0
